requirements.txt
openai
streamlit
langchain-openai
chromadb
langchain-community
PyMuPDF
gtts
python-dotenv